ssi_producer/requirements.txt
kafka-python==2.0.2
websocket-client==1.5.1
